requirements.txt
absl-py==2.1.0
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
bleach==6.1.0
build==1.1.1
CacheControl==0.14.0
cbor2==5.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
cleo==2.1.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
component==0.0.1
constantly==23.10.4
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dm-tree==0.1.8
docstring_parser==0.16
dulwich==0.21.7
efficientnet_pytorch==0.6.3
etils==1.10.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
gymnasium==0.29.1
h11==0.14.0
h5py==3.12.1
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
immutabledict==4.2.0
importlib_resources==6.4.5
incremental==22.10.0
installer==0.7.0
ipykernel==6.29.4
ipython==8.22.2
isoduration==20.11.0
jaraco.classes==3.3.1
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.8.0
joblib==1.3.2
json5==0.9.24
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.6.0
keyring==24.3.1
kiwisolver==1.4.5
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
lazy_loader==0.3
libclang==18.1.1
llama-cloud==0.1.6
llama-index==0.12.4
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.4
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-huggingface-api==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.4.0
llama-index-llms-openai==0.3.3
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-index-readers-web==0.3.1
llama-index-utils-huggingface==0.3.0
llama-parse==0.5.17
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==2.5.0
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.20.0
openai==1.57.0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.4.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pkginfo==1.10.0
platformdirs==4.2.0
playwright==1.49.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
pooch==1.8.1
pretrainedmodels==0.7.4
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==5.28.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.16.1
pyarrow==18.0.0
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyee==12.0.0
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
pyproject_hooks==1.0.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.5.4
pyreadr==0.5.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
rapidfuzz==3.6.2
referencing==0.34.0
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
rpy2==3.5.16
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
segmentation-models-pytorch==0.2.1
selenium==4.27.1
Send2Trash==1.8.2
sentence-transformers==3.3.1
setuptools==69.5.1
sgmllib3k==1.0.0
shellingham==1.5.4
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spider-client==0.0.27
SQLAlchemy==2.0.36
stack-data==0.6.3
striprtf==0.0.26
swig==4.2.1
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.1.2
tensorflow==2.19.0
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
text-generation==0.7.0
tf_keras==2.19.0
threadpoolctl==3.4.0
timm==0.4.12
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.4
torch==2.5.1
torchaudio==2.2.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.46.2
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.3.3
tweepy==4.14.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.0
Wikipedia-API==0.7.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.interface==6.3